requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
gunicorn
google-auth-oauthlib
google-auth
google-api-python-client
itsdangerous
PyJWT
